Make Report an ORDB SubgraphRoot
