+++ examples/demo/top.ord
@@ -8,7 +8,6 @@ from counter_yosys import extlib, report_digital_design
 from ordec.examples.vco_pseudodiff import Vco, VcoHalfStage, VcoRing, VcoTb
 from ordec.layout import helpers
 from ordec.sim import Simulator
-from ordec.report import *
 
 @generate_func
 def report_analog_design():
