+++ ordec/examples/amp.ord
@@ -1,7 +1,6 @@
 from ordec.core import *
 from ordec.lib.base import Res, Cap, Vdc, Vsin, Gnd, Idc
 from ordec.lib.ihp130 import Nmos, Pmos
-from ordec.report import Report
 
 cell Amp:
     viewgen symbol -> Symbol:
