+++ ordec/examples/vco_pseudodiff.ord
@@ -13,7 +13,6 @@ from ordec.core import *
 from ordec.lib.ihp130 import Nmos, Pmos, SG13G2
 from ordec.lib.base import Res, Cap, Gnd, Vdc, NoConn, Vpwl
 from ordec.layout import makevias, SRouter
-from ordec.report import Report
 
 cell VcoHalfStage:
     width = Parameter(R, default=300n)
